+++ tests/test_plot/test_misc.py
@@ -1266,3 +1266,72 @@ def test_dont_stop_1(test_dir):
     ctx.expect_out_file(pos_bot)
     ctx.search_err('ERROR:Failed to create BoM')
     ctx.clean_up(keep_project=True)
+
+
+def test_diff_file_1(test_dir):
+    """ Difference between the current PCB and a reference file """
+    prj = 'light_control_diff'
+    yaml = 'diff_file_'+('k5' if context.ki5() else 'k6')
+    ctx = context.TestContext(test_dir, prj, yaml)
+    ctx.run()
+    ctx.compare_pdf(prj+'-diff.pdf', reference='light_control-diff.pdf')
+    ctx.clean_up(keep_project=True)
+
+
+def test_diff_git_1(test_dir):
+    """ Difference between the current PCB and the git HEAD """
+    prj = 'light_control'
+    yaml = 'diff_git_1'
+    ctx = context.TestContext(test_dir, prj, yaml)
+    # Create a git repo
+    ctx.run_command(['git', 'init', '.'], chdir_out=True)
+    # Copy the "old" file
+    pcb = prj+'.kicad_pcb'
+    file = ctx.get_out_path(pcb)
+    shutil.copy2(ctx.board_file, file)
+    shutil.copy2(ctx.board_file.replace('.kicad_pcb', context.KICAD_SCH_EXT),
+                 file.replace('.kicad_pcb', context.KICAD_SCH_EXT))
+    # Add it to the repo
+    ctx.run_command(['git', 'add', pcb], chdir_out=True)
+    ctx.run_command(['git', 'commit', '-m', 'Reference'], chdir_out=True)
+    # Copy the "new" file
+    shutil.copy2(ctx.board_file.replace(prj, prj+'_diff'), file)
+    # Run the test
+    ctx.run(extra=['-b', file], no_board_file=True)
+    ctx.compare_pdf(prj+'-diff.pdf')
+    ctx.clean_up(keep_project=True)
+
+
+def test_diff_git_2(test_dir):
+    """ Difference between the two repo points """
+    prj = 'light_control'
+    yaml = 'diff_git_2'
+    ctx = context.TestContext(test_dir, prj, yaml)
+    # Create a git repo
+    ctx.run_command(['git', 'init', '.'], chdir_out=True)
+    # Copy the "old" file
+    pcb = prj+'.kicad_pcb'
+    file = ctx.get_out_path(pcb)
+    shutil.copy2(ctx.board_file, file)
+    shutil.copy2(ctx.board_file.replace('.kicad_pcb', context.KICAD_SCH_EXT),
+                 file.replace('.kicad_pcb', context.KICAD_SCH_EXT))
+    # Add it to the repo
+    ctx.run_command(['git', 'add', pcb], chdir_out=True)
+    ctx.run_command(['git', 'commit', '-m', 'Reference'], chdir_out=True)
+    # Add an extra commit
+    dummy = ctx.get_out_path('dummy')
+    with open(dummy, 'wt') as f:
+        f.write('dummy\n')
+    ctx.run_command(['git', 'add', 'dummy'], chdir_out=True)
+    ctx.run_command(['git', 'commit', '-m', 'Dummy noise'], chdir_out=True)
+    # Copy the "new" file
+    shutil.copy2(ctx.board_file.replace(prj, prj+'_diff'), file)
+    # Add it to the repo
+    ctx.run_command(['git', 'add', pcb], chdir_out=True)
+    ctx.run_command(['git', 'commit', '-m', 'New version'], chdir_out=True)
+    # Now just wipe the current file
+    shutil.copy2(ctx.board_file.replace(prj, '3Rs'), file)
+    # Run the test
+    ctx.run(extra=['-b', file], no_board_file=True, extra_debug=True)
+    ctx.compare_pdf(prj+'-diff.pdf')
+    ctx.clean_up(keep_project=True)